fixed insuficientstockerror result

+++ api-gateway/src/graphql/schema/types/CheckoutResultType.ts
@@ -3,7 +3,7 @@ import { gql } from "apollo-server";
 const CheckoutResultType = gql`
   type CheckoutResult {
     status: String!
-    itemsWithInsuficientStock: [Int]
+    itemsWithInsuficientStock: [Long]
   }
 `;
